requirements.txt
# Core dependencies - updated to latest stable versions (Oct 2025)
chromadb>=1.2.0,<2.0.0     # Vector database (upgraded from 0.4.22)
sentence-transformers>=5.1.0,<6.0.0  # Embeddings (upgraded from 2.3.1)
watchdog==3.0.0
pyyaml==6.0.1
requests==2.31.0
numpy<2.0.0

# Phase 1.5: LevelDB parser for Claude Desktop
plyvel==1.5.0              # LevelDB Python bindings

# Phase 2: Embedding & LLM Integration
torch>=2.1.0               # PyTorch for sentence-transformers
